Call ensureOVNMetricsCertIfNeeded also during minor updates
During a minor update we call ReconcileOVNController and not
ReconcileOVN, so if metrics are enabled we would hit an issue where the
necessary TLS certificate will not be created and reconciliation would
get stuck.
With this change we move ensureOVNMetricsCertIfNeeded to its own
function so that it can be called from both ReconcileOVNController and
ReconcileOVN as necessary.